requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
Jinja2==3.1.4
Flask-Session==0.7.0
pydantic==2.8.2
scikit-learn==1.5.1
scipy==1.13.1
numpy==1.26.4
spacy==3.7.5
nltk==3.9.1
sumy==0.11.0
python-docx==1.1.2
reportlab==4.2.2
itsdangerous==2.2.0
click==8.1.7